Moved some methods into lisp-objects.h/c
