+++ include/math.h
@@ -18,44 +18,6 @@
 obj* get_math_library();
 
 /**
- * Function: new_int
- * -----------------
- * Creates a new integer object wrapping a raw integer value.
- * Note: This object will not be automatically added to the "allocated" list
- * @param value: The integer value to wrap in an object
- * @return: The object wrapping a copy of the integer value in dynamically allocated memory
- */
-obj* new_int(int value);
-
-/**
- * Function: new_float
- * -------------------
- * Creates a new float object wrapping a raw integer value.
- * Note: This object will not be automatically added to the "allocated" list
- * @param value: The float value to wrap in an object
- * @return: The object wrapping a copy of the floating point value
- */
-obj* new_float(float value);
-
-/**
- * Function: get_int
- * -----------------
- * Get the integer value from an object wrapping a number
- * @param o: The object that is wrapping the integer or the float
- * @return: A copy of the value stored in the object as an integer
- */
-int get_int(const obj* o);
-
-/**
- * Function: get_float
- * -------------------
- * Get the floating point value from an object wrapping a number
- * @param o: The object wrapping the integer or floating point value
- * @return: A copy of the value stored in the object as a floating point
- */
-float get_float(const obj* o);
-
-/**
  * Primitive: plus
  * ---------------
  * Primitive function for adding two numbers
